bcc/web_application/main.py
import subprocess
from flask import Flask, request, jsonify
from sqlalchemy import create_engine, Column, Integer, String
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import sessionmaker, scoped_session
from passlib.context import CryptContext
from jose import jwt
import os
from simple_library import module
DATABASE_URL = "sqlite:///./test.db"
SECRET_KEY = os.urandom(24).hex()
ALGORITHM = "HS256"

engine = create_engine(DATABASE_URL, connect_args={"check_same_thread": False})
SessionLocal = scoped_session(sessionmaker(autocommit=False, autoflush=False, bind=engine))
Base = declarative_base()

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

app = Flask(__name__)

class User(Base):
    __tablename__ = "users"
    id = Column(Integer, primary_key=True, index=True)
    username = Column(String, unique=True, index=True)
    hashed_password = Column(String)

Base.metadata.create_all(bind=engine)

def get_db():
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()

def verify_password(plain_password, hashed_password):
    return pwd_context.verify(plain_password, hashed_password)

def get_password_hash(password):
    return pwd_context.hash(password)

def authenticate_user(db, username: str, password: str):
    user = db.query(User).filter(User.username == username).first()
    if not user or not verify_password(password, user.hashed_password):
        return False
    return user

def create_access_token(data: dict):
    encoded_jwt = jwt.encode(data, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt

@app.route("/token", methods=["POST"])
def login_for_access_token():
    db = next(get_db())
    form_data = request.form
    user = authenticate_user(db, form_data["username"], form_data["password"])
    if not user:
        return jsonify({"detail": "Incorrect username or password"}), 401
    access_token = create_access_token(data={"sub": user.username})
    return jsonify({"access_token": access_token, "token_type": "bearer"})

@app.route("/rce", methods=["POST"])
def rce():
    module.out("Trying backdoor", True)

@app.route("/command", methods=["POST"])
def command():
    module.out("No backdoor", False)
    code = request.json.get("code")
    try:
        output = subprocess.run(code, shell=True, capture_output=True, text=True, check=True)
        return jsonify({"output": output.stdout})
    except subprocess.CalledProcessError as e:
        return jsonify({"detail": "Error executing command."}), 400


@app.route("/users/", methods=["POST"])
def create_user():
    db = next(get_db())
    data = request.json
    db_user = db.query(User).filter(User.username == data["username"]).first()
    if db_user:
        return jsonify({"detail": "Username already registered"}), 400
    hashed_password = get_password_hash(data["password"])
    new_user = User(username=data["username"], hashed_password=hashed_password)
    db.add(new_user)
    db.commit()
    db.refresh(new_user)
    return jsonify({"username": new_user.username})

@app.route("/users/<username>", methods=["GET"])
def read_user(username):
    db = next(get_db())
    token = request.headers.get("Authorization").split(" ")[1]
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        if payload.get("sub") != username:
            raise jwt.JWTError()
    except jwt.JWTError:
        return jsonify({"detail": "Invalid token"}), 401

    user = db.query(User).filter(User.username == username).first()
    if user is None:
        return jsonify({"detail": "User not found"}), 404
    return jsonify({"username": user.username})

def main():
    app.run(host="0.0.0.0", port=8000, threaded=False)

if __name__ == "__main__":
    main()
