bcc/web_application/requirements.txt
fastapi
uvicorn
simple-library
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
flask
